refactor: :recycle: refactor all components to jsx
