+++ my-app/src/Routes.js
@@ -3,11 +3,11 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import "./App.css";
 import Home from "./pages/Home";
 import Members from "./pages/Members";
-// import Event from "./core/Event";
+import Event from "./pages/Event";
 import Contact from "./pages/Contact";
 import About from "./pages/About";
-// import Error from "./core/Error";
-// import Alumni from "./core/Alumni";
+import Error from "./pages/404";
+import Alumni from "./pages/Alumni.jsx";
 
 const Routes = () => {
   return (
@@ -15,13 +15,13 @@ const Routes = () => {
       <Router>
         <Switch>
           <Route path="/" exact component={Home} />
-          <Route path='/members' exact component={Members} />
-          <Route path='/about' exact component={About} />
-          <Route path='/contact' exact component={Contact} />
-          {/* <Route path='/alumni' exact component={Alumni} />
-          <Route path='/events' exact component={Event} />
-          <Route path='/contact' exact component={Contact} />
-          <Route path='' component={Error} /> */}
+          <Route path="/members" exact component={Members} />
+          <Route path="/alumni" exact component={Alumni} />
+          <Route path="/about" exact component={About} />
+          <Route path="/contact" exact component={Contact} />
+          <Route path="/events" exact component={Event} />
+
+          <Route path="" component={Error} />
         </Switch>
       </Router>
     </div>
